Added display 'output' numerical/identifier prefix
On some systems (like as a VirtualBox guest), the display-settings
dialog needs better, discerning identifiers for displays.  It is
much easier to find a display when it is numbered than when all
displays share the same friendly name because they are all from
the same vendor.

This fix prefixes the 'output' number to the friendly-but-not-unique
names.  It should apply to 3 main UI elements:

- List of displays
- Dropdown for selecting display position
- Display identification "floater" boxes (when "Identify Displays"
  button is pressed).